基本调试完成，但频率计算有误差

+++ User/OSC/OSC.h
@@ -20,13 +20,13 @@ extern  char*     CurSamplStatus;   //代号0，采样状态，0：停止采样，1：正在采样
 extern  float     CurTriggerValue;  //代号1，触发阀值
 extern  char*     CurTriggerMode;   //代号2，触发模式，0：下降沿触发，1：上升沿触发，2：上升沿下降沿触发
 extern  char*     CurSamplingMode;  //代号3，采样模式，0：自动，1：普通，2：单次
-extern  uint16_t  CurTimePerDiv;    //代号4，每格代表的时间间隔
+extern  uint32_t  CurTimePerDiv;    //代号4，每格代表的时间间隔
 
 //要显示的信息
 extern __IO        uint16_t     ADC_ConvertedValue[];//ADC采集数据
-extern             float        CurWaveFrq;//波形频率，单位kHz
+extern             double        CurWaveFrq;//波形频率，单位kHz
 
-extern 						 uint16_t     TimePerDiv_Group[];
+extern 						 uint32_t     TimePerDiv_Group[];
 extern             uint8_t     	TimePerDivOderNbr;
 extern             int8_t     	TimePerDivOder;//当前每格间隔时间的序号
 extern             char*        SamplStatus[];
